Rename dna_optimiser to generic subst_optimiser
